Added basic contribution guidelines
Getting started with making a contribution to the project. A how-to with a few bits of advice on making a good PR. Tried to make the description friendly to new oss contributors...considering I'm part of that group.

Thanks and credit goes to @TalAter. I referenced his post (here)[https://github.com/TalAter/Progressive-UI-KITT/issues/1] to make the guidelines.